rewriting calculator via class
not using GUI yet

+++ python/operations.py
@@ -1,13 +1,13 @@
 def add(x, y):
-    return float(int(x) + int(y))
+    return float((x) + (y))
 
 def subtract(x, y):
-    return float(int(x) - int(y))
+    return float((x) - (y))
 
 def multiply(x, y):
-    return float(int(x) * int(y))
+    return float((x) * (y))
 
 def divide(x, y):
-    if int(y) == 0:
+    if float(y) == 0.0:
         print("Can't be")
-    return float(int(x) / int(y))
+    return float((x) / (y))
